requirements.txt
appnope==0.1.0
birdy==0.2
decorator==4.0.9
gnureadline==6.3.3
ipython==4.1.2
ipython-genutils==0.1.0
kafka-python==1.0.2
oauthlib==1.0.3
path.py==8.1.2
pexpect==4.0.1
pickleshare==0.6
ptyprocess==0.5.1
PyYAML==3.11
requests==2.9.1
requests-oauthlib==0.6.1
simplegeneric==0.8.1
six==1.10.0
traitlets==4.2.1
